+++ package.json
@@ -16,9 +16,9 @@
     "aos": "^2.3.4",
     "eslint": "8.34.0",
     "eslint-config-next": "13.1.6",
-    "next": "13.1.6",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "next": "^14.2.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "typescript": "4.9.5"
   },
   "devDependencies": {
